updated list resources options + paginated results + move types to type.ts
