+++ nexus-sdk-js/src/Resource/index.ts
@@ -1,4 +1,4 @@
-import { Context } from './types';
+import { Context, ResourceResponse } from './types';
 import {
   getResource,
   getSelfResource,
@@ -28,32 +28,6 @@ export const RESOURCE_METADATA_KEYS = [
   '_rev',
   '_deprecated',
 ];
-
-export interface ResourceResponseCommon {
-  '@id': string;
-  '@type'?: string | string[];
-  _self: string;
-  _constrainedBy: string;
-  _project: string;
-  _createdAt: string;
-  _createdBy: string;
-  _updatedAt: string;
-  _updatedBy: string;
-  _rev: number;
-  _deprecated: boolean;
-  // This is for the rest of the dataset's data
-  [key: string]: any;
-}
-
-export interface ListResourceResponse {
-  '@context'?: Context;
-  _total: number;
-  _results: ResourceResponseCommon[];
-}
-
-export interface ResourceResponse extends ResourceResponseCommon {
-  '@context'?: Context;
-}
 
 export default class Resource<T = {}> {
   readonly orgLabel: string;
